only show login button when not logged in
